core: integrate new schema API
this still has one failing test. incoming records with an unknown type
currently fail silently (which is bad). the schema.Record function
throws on unknown types. Instead, it should return a MissingRecord
record with most functionality still working.

This still points at a bigger problem though: we want types and their
schemas before we index records. For this easiest for now will be to use
a seperate feed type.