core, db: have two feeds per collection (root and data)
This change introduces a seperate root feed for a collection. Only Feed
and Type records are put into that feed. All other records end up in a
data feed. The data feed is putted into the root feed as a Feed record.

This is so we can have a new root kappa view that is scoped to only
the root feeds. This allows to wait for the root kappa to complete sync
before throwing for schema errors.