Add authentication to the HTTP API.
This requires a valid JSON web token in the Authorization header for all
API requests. Tokens can be generated through HTTP requests when
providing an access code. Access codes can be created through HTTP
requests by an initial access code, which is available to the Sonar UI
on the same hosts, and can be passed to browsers through login links.

TODO:

- Support session keys to reduce overhead of passing JWTs in each
  request
- Add capabilities to individual collections (either read or write) to
  the JWTs (derived from the Hypercore crypto)
- Support multiple JWTs per session