switch to fetch-undici
this is the new fetch implementation in Node.js that is far more spec
compliant than node-fetch, which we were using before.

the most notable change is that we now always return Uint8Arrays if
responseType is set to buffer. Before, that'd return Uint8Array in
browsers and Buffer in Node.js.

also fixes typescript typings and file tests for this.